From ghaskins@novell.com Sat May 24 00:14:29 2008 To: mingo@elte.hu, tglx@linutronix.de, rostedt@goodmis.org,      linux-rt-users@vger.kernel.org Cc: linux-kernel@vger.kernel.org, sdietrich@novell.com, pmorreale@novell.com,      mkohari@novell.com, ghaskins@novell.com Subject: [PATCH 5/5] remove the extra call to try_to_take_lock
    [ The following text is in the "utf-8" character set. ]
    [ Your display is set for the "iso-8859-1" character set.  ]
    [ Some characters may be displayed incorrectly. ]

From: Peter W. Morreale <pmorreale@novell.com>

Remove the redundant attempt to get the lock.  While it is true that the
exit path with this patch adds an un-necessary xchg (in the event the
lock is granted without further traversal in the loop) experimentation
shows that we almost never encounter this situation.

Signed-off-by: Peter W. Morreale <pmorreale@novell.com>
Signed-off-by: Gregory Haskins <ghaskins@novell.com>