From ghaskins@novell.com Tue May 27 21:24:29 2008 To: Steven Rostedt <rostedt@goodmis.org>, linux-rt-users@vger.kernel.org Cc: linux-kernel@vger.kernel.org, Ingo Molnar <mingo@elte.hu>,      Thomas Gleixner <tglx@linutronix.de>,      Gregory Haskins <ghaskins@novell.com> Subject: [PATCH 1/3] sched: fix RT task-wakeup logic
    [ The following text is in the "utf-8" character set. ]
    [ Your display is set for the "iso-8859-1" character set.  ]
    [ Some characters may be displayed incorrectly. ]

Dmitry Adamushko pointed out a logic error in task_wake_up_rt() where we
will always evaluate to "true".  You can find the thread here:

http://lkml.org/lkml/2008/4/22/296

In reality, we only want to try to push tasks away when a wake up request is
not going to preempt the current task.  So lets fix it.

Note: We introduce test_tsk_need_resched() instead of open-coding the flag
check so that the merge-conflict with -rt should help remind us that we
may need to support NEEDS_RESCHED_DELAYED in the future, too.

Signed-off-by: Gregory Haskins <ghaskins@novell.com>
CC: Dmitry Adamushko <dmitry.adamushko@gmail.com>
CC: Steven Rostedt <rostedt@goodmis.org>
Signed-off-by: Ingo Molnar <mingo@elte.hu>