Subject: rtmutex: check integrity
When PREEMPT_RT is configured on, a spinlock or semaphore can turn into
a rt_mutex. Since they also may stay the same in some cases (depending on
what type they were defined as) the API for them is determined by the
type. For example, if a spin lock is of type spinlock_t it will be converted
to an rt_mutex, and if it is defined as raw_spinlock_t it will stay the
same.

If the locks are passed as pointers and typecasted to the wrong lock,
things can break. This patch adds a integrity check to make sure that
the rt_mutexs are indeed rt_mutexes when used.

Signed-off-by: Steven Rostedt <srostedt@redhat.com>