Subject: ftrace: fix function profiler race
fix function profiler race

The bug was caused by a race in the code. The profile function could have been
preempted just before disabling interrupts. Then after the trace is disabled,
and freed, and then started again, the function could continue and use
uninitialized pointers, or corrupted pointers.

This patch adds a check inside the irq disabled section to see if the profile
is active or not. The registering and unregistering of the trace function will
act like a barrier since it calls stop machine.

Signed-off-by: Steven Rostedt <srostedt@redhat.com>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>