Subject: fix printk in atomic hack
The printk in atomic hack had a slight bug. This but was triggered
by debug locking options. The hack prevents grabbing sleeping spin
locks in printk console drivers if we are in atomic (can't sleep).
But the unlock had a bug where it incorrectely assumed that if
we are in printk and atomic, that we didn't grab the lock. The debug
locking can encapsulate these options and cause unlocks to be in atomic
when the lock was not. This means we would not release the lock after
it was taken.

The patch only skips releasing the lock if in printk - atomic *and*
not the lock owner.

Special thanks goes to Jon Masters for digging his head deep into
this crap and narrowing it down to a problem with printks and locks.

Signed-off-by: Steven Rostedt <srostedt@redhat.com>
Signed-off-by: Ingo Molnar <mingo@elte.hu>