Increase hung_task_timeout_secs defaul on RT
The default value of the task hangcheck timer is hardcoded to 120 seconds.
This default is too low for RT environments. As a test measure we are
increasing this value to 600s (10 minutes).

The default value lead to several test abortions on RHTS due to kernel panics.
Meaning that a task (several cases had yum-updatesd) was hung for 120s during a
stress test - except by the kernel panic, the rest was expected.

Signed-off-by: Luis Claudio R. Goncalves <lgoncalv@redhat.com>