drm/nouveau: change MMIO map to simple ioremap
This removes the ugliness of using drm_local_map. The use of ioremap()
is identical to what drm_addmap()/drm_rmmap() did.

Is ioremap() the right thing or should we use e.g. ioremap_uc()?
The ioremap API has been reworked in the past, and some assumptions
changed in the mapping caching type.

Signed-off-by: Pekka Paalanen <pq@iki.fi>