drm/nouveau: io{read,write}32_native API
Try to clean up the register accessor definitions by defining
io{read,write}32_native API, which does not byteswap, and rewrite all
accessors to use it.

As a consequence, readl/writel calls are changed to ioread32/iowrite32
also on x86. It adds a tiny overhead, since they are generic IO
functions, but the overhead should be diminishing as compared to access
over PCI bus.

All 16-bit accessors are removed as unused.

Signed-off-by: Pekka Paalanen <pq@iki.fi>