drm/nouveau: bo read/write wrappers for nv04_crtc.c
Introduce accessors for TTM buffer object memory that has been mapped
into the kernel virtual address space or as IO memory. IO memory needs
to be accessed via special accessor functions, not by dereferencing the
iomem cookie. The wrappers hide the details of 32-bit access and honour
the TTM map type.

nv04_crtc_cursor_set() is changed to use the new wrappers. 'cursor' is
received from user space, which means we cannot assume that it is or is
not iomem. 'nv_crtc->cursor.nvbo' should always be iomem, and as such
the access is relatively slow. Therefore using the wrappers for it, too,
should not cost much.

Signed-off-by: Pekka Paalanen <pq@iki.fi>
Signed-off-by: Ben Skeggs <bskeggs@redhat.com>