drm/nouveau: remove mask in nv04_fbcon_imageblit()
The local variable 'mask' is simply confusing, the value is equivalent
to zero.

Why? The literal 0 is by default signed. So is ~0. Bitshift to right of
a signed type is implementation defined, gcc replicates the sign bit.
Therefore ~0 >> x equals always ~0, and the final 'mask' value is 0.

Remove this obfuscation.

Signed-off-by: Pekka Paalanen <pq@iki.fi>
Signed-off-by: Ben Skeggs <bskeggs@redhat.com>