drm/nv04: fix null pointer dereferences of native_mode
nv_connector->native_mode is not set when no modes were found for the
connector, so its existence can't be assumed.

In nv04_dfp_mode_fixup, reject the mode if GPU scaling is enabled but
native mode is not known.

In nv04_dfp_mode_set, use clock value from output_mode (nv_encoder->mode)
instead of nv_connector->native_mode. If panel scaling is enabled on a
TMDS display and the display did not have a valid EDID, native_mode is
NULL. Also, the use of native_mode clock here seems wrong, as a dual link
monitor is not always operated in dual link mode.

In nv04_lvds_dpms and nv04_dfp_restore, refuse to turn on an LVDS panel
if native mode is not known. While clock is not always required for
turning panel on, having an LVDS without native mode means something went
wrong already, so trying to turn panel on only in cases where clock is
not required would probably not be worth it.

This fixes http://bugs.freedesktop.org/show_bug.cgi?id=23295

Signed-off-by: Anssi Hannula <anssi.hannula@iki.fi>