drm/nv50: implement pgraph.unload_context()
The semantics of this new hook are slightly different to the existing
pgraph.save_context().  The former would save the current state of
PGRAPH into "chan", and do nothing further.  This call saves the current
state of PGRAPH into the channel that's currently active on PGRAPH,
and then invalidates the context on PGRAPH.