smi-remediation-simple-fix
smi-remediation-simple-fix

As part of validing 2.6.31 kernel I fixed a know problem with the i5000_edac
driver.   In some
topolgies the driver would fail to expose memory devices via sysfs and the
ibm-prtmd service would
not start.

It is a pretty simple fix and it has one Ack so far on the edac mainling list.
I don't see why it will not go upstream.

Without this patch smi remediation  is not enable on about 1/2 of the possible
memory topologies.

Signed-off-by: Darren Hart <dvhltc@us.ibm.com>