drm/nouveau: rework bios delays
Replace the BIOS_USLEEP() macro with appropriate calls to udelay() and
msleep(). BIOS_USLEEP() was implemented with mdelay(), which is a
busy-wait. msleep() on the other hand can reschedule, so these paths
cannot be called from interrupt or atomic context.

In init_reset(), BIOS_USLEEP(10) was a no-op, since the delay was
rounded off to zero. This was probably a bug in porting the DDX code to
the kernel. The delay is restored.

In init_time(), since the delay is variable with a large range, it is
implemented as a busy-wait for <1ms, and msleep() otherwise rounding
mostly to the next integer milliseconds.

call_lvds_manufacturer_script() contained a bug in the wait: the delay
was executed in microseconds, when it should be milliseconds. Fixed.

Signed-off-by: Pekka Paalanen <pq@iki.fi>