From 0cfa166f6f03cf44439f0ea7db5278d1438f017c Mon Sep 17 00:00:00 2001 Subject: [PATCH] Remove pulse code from the bnx2x driver
We are still seeing pulse mismatch error messages while running
daEth on the receiver.  While not sure why my development testing
did not reveal this, release testing did.  Here is another go that
should fix it for sure.

Reasoning for this is from the driver maintainer, Eilon Greenstein:
A possible work around is to remove the pulse section from the
timer - everything will work just fine without it. It is there
to allow the FW to take control over the link incase the driver
died unexpectedly, but since you are not using NCSI (or any other
management protocol on this interface) - you don't really need it.

Built and tested against 2.6.31-rc8-rt9 kernel.

Signed-off-by: Vernon Mauery <vernux@us.ibm.com>