drm/nouveau: refuse to load on un-posted chipsets
We can POST G8x chips just fine, but not earlier chips due to the reliance
on VBIOS initialised values for the INIT_COMPUTE_MEM opcode.

From looking at various kmmio traces it appears NVIDIA check for a POSTed
chipset the same way as on G8x, so we'll assume that's correct.  Hit me
up on IRC if nouveau wrongly refuses to load for you now!

Signed-off-by: Ben Skeggs <bskeggs@redhat.com>