perf_events: fix errors path in perf_output_begin() (take 3)
In case the sampling buffer has no "payload" pages, nr_pages is 0.
The problem is that the error path in perf_output_begin() skips to
a label which assumes perf_output_lock() has been issued which is
not the case. That triggers a WARN_ON() in perf_output_unlock().

This patch fixes the problem by skipping perf_output_unlock()
in case data->nr_pages is 0.

Signed-off-by: Stephane Eranian <eranian@google.com>

--