Make sure d_kill sets d_parent to null
Patch against currrent 2.6.33-rt tree.

This patch is an alternative fix to the select_parent panic that was
suggested by Nick Piggin.

It resolves the issue by making sure d_parent is set to null so that
the dentry ascending code in select_parent will restart if the child
or next dentry was dkilled while the locks were released.

Credit and thanks to Nick for finding the solution.

Signed-off-by: John Stultz <johnstul@us.ibm.com>
CC: Thomas Gleixner <tglx@linutronix.de>
CC: Nick Piggin <npiggin@suse.de>
CC: Peter Zijlstra <peterz@infradead.org>
CC: John Kacur <jkacur@redhat.com>