bnx2x: Fix PHY locking problem
commit 971a230f9573cca1cbef96dab05e2682820f1a0 upstream

PHY locking is required between two ports for some external PHYs. Since
initialization was done in the common init function (called only on the
first port initialization) rather than in the port init function, there
was in fact no PHY locking between the ports.

Signed-off-by: Yaniv Rosner <yanivr@broadcom.com>
Signed-off-by: Eilon Greenstein <eilong@broadcom.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Cherry-picked-for: 2.6.33.9
(Fixed patch to reflect the new location of the file)
Signed-off-by: John Kacur <jkacur@redhat.com>