mm: make the vma list be doubly linked
commit 297c5eee372478fc32fec5fe8eed711eedb13f3d upstream

It's a really simple list, and several of the users want to go backwards
in it to find the previous vma.  So rather than have to look up the
previous entry with 'find_vma_prev()' or something similar, just make it
doubly linked instead.

Tested-by: Ian Campbell <ijc@hellion.org.uk>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Cherry-picked for v2.6.33.7-rt29
Resolved conflict in kernel/fork.c
Signed-off-by: John Kacur <jkacur@redhat.com>